requirements.txt
APScheduler==3.10.1
discord.py==2.3.2
imagehash==4.3.1
mysql-connector-python==8.1.0
paramiko==3.4.0
Pillow==10.1.0
python-dotenv==1.0.1
pytz==2023.3
requests==2.31.0
validators==0.22.0
pyinstaller==6.1.0
PyDrive==1.3.1